fix some of cooks assistant quest
